requirements.txt
annotated-types==0.7.0
attrs==25.1.0
boto3==1.36.11
botocore==1.36.11
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.1
execnet==2.1.1
fonttools==4.55.8
idna==3.7
importlib_resources==6.5.2
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.4
motor==3.7.0
mysql-connector-python==9.2.0
numpy==2.0.2
packaging==24.0
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.2.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-on-whales==0.75.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
zipp==3.21.0
